+++ src/surrealist/ql/statements/insert_statements.py
@@ -2,6 +2,13 @@ from surrealist.ql.statements.statement import FinishedStatement, Statement
 
 
 class OnDuplicate(FinishedStatement):
+    """
+    Represents ON DUPLICATE part of the INSERT statement
+
+    [ ON DUPLICATE KEY UPDATE @field = @value ... ]
+    Refer to:
+    https://surrealdb.com/docs/surrealdb/surrealql/statements/insert
+    """
     def __init__(self, statement: Statement, action: str):
         super().__init__(statement)
         self._action = action
@@ -12,4 +19,12 @@ class OnDuplicate(FinishedStatement):
 
 class InsertUseDuplicate:
     def on_duplicate(self, action: str) -> OnDuplicate:
+        """
+        Add ON DUPLICATE clause to final statement
+
+        ON DUPLICATE KEY UPDATE @field = @value ...
+
+        :param action: raw representation of action
+        :return: OnDuplicate object
+        """
         return OnDuplicate(self, action)